Lazy boxed: add Unsafe modules
