+++ vector/tests/doctests.hs
@@ -26,6 +26,8 @@ main = mapM_ run modGroups
         ]
       , [ "src/Data/Vector.hs"
         , "src/Data/Vector/Mutable.hs"
+        , "src/Data/Vector/Unsafe.hs"
+        , "src/Data/Vector/Mutable/Unsafe.hs"
         ]
       , [ "src/Data/Vector/Strict.hs"
         , "src/Data/Vector/Strict/Mutable.hs"
